[REFACTOR] GuestRepository 패키지 수정 (#153)
