+++ src/main/java/org/chzzk/howmeet/domain/temporary/guest/repository/GuestRepository.java
@@ -1,7 +1,7 @@
-package org.chzzk.howmeet.domain.temporary.auth.repository;
+package org.chzzk.howmeet.domain.temporary.guest.repository;
 
 import org.chzzk.howmeet.domain.common.model.Nickname;
-import org.chzzk.howmeet.domain.temporary.auth.entity.Guest;
+import org.chzzk.howmeet.domain.temporary.guest.entity.Guest;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.data.repository.query.Param;
