+++ src/test/java/org/chzzk/howmeet/domain/temporary/auth/repository/GuestRepositoryTest.java
@@ -1,7 +1,8 @@
 package org.chzzk.howmeet.domain.temporary.auth.repository;
 
 import org.chzzk.howmeet.domain.common.model.Nickname;
-import org.chzzk.howmeet.domain.temporary.auth.entity.Guest;
+import org.chzzk.howmeet.domain.temporary.guest.entity.Guest;
+import org.chzzk.howmeet.domain.temporary.guest.repository.GuestRepository;
 import org.chzzk.howmeet.global.config.RepositoryTest;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.params.ParameterizedTest;
